==> 30.Figure_3_and_Supplementary_Fig5-6/plot_Fig3g-h_and_Supp_Fig_5b-d/12.getPhenoDat.R <==
library(tidyverse)
library(magrittr)
library(rjson)

params <- fromJSON(file = "../00.config/config.json")
wells <- file.path(params$analysisDir, params$listOfWells) %>%
        readLines()

load("barcodes_to_cells.rds")
load("features.rds")
phenoDat <- merge(features, bcdCalls[, c("bcd", "objId", "sgId", "gene", "poolId")],
                by.x = "label", by.y = "objId")

save(phenoDat, file = "phenotypic_data.rds")
save(phenoDat, file = "normalized_phenotypic_data.rds")
